+++ src/commands/mod.rs
@@ -5,4 +5,5 @@
 pub mod about;
 pub mod join;
 pub mod play;
+pub mod roll;
 pub mod seek;